val_dataloader defined, it will consits of only a pair (input_seed, target).
Modified the on_before_batch_transfer hook:
now, before a batch is transferred from cpu to gpu, the first element
of the batch will still be the input_seed, but the remaining elements of
that batch will be cropped with the function take_cube.

Two functions defined here:
1) make_cube_damage, which replicates the distill-paper strategy of square damage
to a (previously sampled from pool) pattern. But, using this, then there will be
some problems in damaging A LOT the grown object, so I wrote it but can be enanched
(just my opinion)

2) take_cube, which is an enhanced and more general version/idea of make_cube_damage.
In take_cube, from each developed pattern (sampled from pool) a region is taken and used as
new input, at the spot of the pattern just damaged (as make_cube_damage and the
distill-paper authors do).
In this way, the model understands how to regenerate from some damage (same as the paper)
but even if the damage is enormous it can restore the entire object from it. In fact,
just reproducing the paper idea (using make_cube_damage), some enormous damage like the
70-90% of the object will not allow to reproduce the entire object again, sometimes it will also collapse
to itself until every single pixel disappear.
Other than this, the original paper idea was very dependent from the starting seed: when using a seed in a
different spot, the model will not work at all!
So, this take_cube function will help the model to be independent to the starting position of the seed
because each time he need to recover the entire object from a cube-shape box with limited amount of pixels.