make_cube_damage and take_cube are nearly the same of those functions in VoxelDataModule with just a few changes due to the fact that here will be used only on batches with len 1.
validation_step defined, logging loss_from_seed and loss_reconstructed. The last one
is used to understand how the model is recovering from huge damages or from starting to
regions unexplored by the starting seed in the first steps.

training_step modifications:
now the model is used two times on two inputs. The first one is the seed, the second
is the previously computed output (first input) which will be damaged with the take_cube function.
Those two outputs will then be used to calculate a loss which is the average of the seed_loss and
reconstruction_loss. In this way, the model is FORCED to understand how to recover from every type and
amount of damage.

in the __init__ method (constructor) weight_norm is used, and kernel_x/y/z are now used as attribute of the
model thanks to the self.register_buffer method (this will avoid unnecessary computation of kernels at each update
step of each train step of each epoch.. which is quite a lot of useless computation).